add case ULONG_MAX

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -10,7 +10,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int bits = 0, var;
+	unsigned long int bits = 0, var;
 
 	var = n ^ m;
 	while (var > 0)
